simplify and reduce features, cfg
