+++ miniconf/tests/packed.rs
@@ -1,5 +1,3 @@
-#![cfg(all(feature = "json-core", feature = "derive"))]
-
 use miniconf::{Packed, Traversal, Tree, TreeKey, TreeSerialize};
 
 #[derive(Tree, Default)]
